fixed #145 replace hammer.js

+++ src/app/client/pages/canvas/draw/components/board.vue
@@ -483,7 +483,9 @@ export default {
       !opid && this.socket.emit('sync', 'undo', item, this.board._id, this.board._id)
     },
     deleteSelected() {
-      this.drawer.deleteSelected()
+      if (this.canDelete) {
+        this.drawer.deleteSelected()
+      }
     },
     choose(chooseKey, hiddenAction) {
       if (!this.plugins[chooseKey].useInFollowing && this.notPresenter) {
